+++ src/headers/mod.rs
@@ -108,7 +108,9 @@ impl MobiMetadata {
     /// Returns raw ExthRecord data located at appropriate position if it exists. It is
     /// highly recommended to use public api provided here to access those records but
     /// in case where lower level access is needed this method provides access to all fields.
-    pub fn exth_record(&self, record: ExthRecord) -> Option<&Vec<u8>> {
+    ///
+    /// Some records can occur multiple times thats why a list of data buffers is returned.
+    pub fn exth_record(&self, record: ExthRecord) -> Option<&Vec<Vec<u8>>> {
         self.exth.get_record(record)
     }
 
@@ -120,7 +122,9 @@ impl MobiMetadata {
     /// It is highly recommended to instead use public api provided here to access
     /// those records but in case where lower level access is needed this method
     /// provides access to all fields.
-    pub fn exth_record_at(&self, position: u32) -> Option<&Vec<u8>> {
+    ///
+    /// Some records can occur multiple times thats why a list of data buffers is returned.
+    pub fn exth_record_at(&self, position: u32) -> Option<&Vec<Vec<u8>>> {
         self.exth.get_record_position(position)
     }